+++ SwiftUI Landmarks/SupportingView/Badge.swift
@@ -2,64 +2,39 @@
 //  Badge.swift
 //  SwiftUI Landmarks
 //
-//  Created by 平林宏淳 on 2020/12/22.
+//  Created by 平林宏淳 on 2020/12/23.
 //  Copyright © 2020 Hiroaki_Hirabayashi. All rights reserved.
 //
 
 import SwiftUI
 
 struct Badge: View {
-    var body: some View {
-       
-        
-        GeometryReader { geometry in
-            Path { path in
-                var width: CGFloat = min(geometry.size.width, geometry.size.height)
-     //六角形の大きさ
-                let height = width
-                let xScale: CGFloat = 0.832
-                let xOffset = (width * (1.0 - xScale)) / 2.0
-                width *= xScale
-
-                path.move(
-                    to: CGPoint(
-                        x: xOffset + width * 0.95,
-                        y: height * (0.20 + HexagonParameters.adjustment) //線画の開始地点
-                    )
-                )
-                
-                HexagonParameters.segments.forEach { segment in
-                    path.addLine(
-                        to: .init(
-                            x: width * segment.line.x + xOffset,
-                            y: height * segment.line.y
-                        )
-                    )
-                    
-                    path.addQuadCurve(
-                    to: .init(
-                        x: width * segment.curve.x + xOffset,
-                        y: height * segment.curve.y
-                    ),
-                    control: .init(
-                        x: width * segment.control.x + xOffset,
-                        y: height * segment.control.y
-                        )
-                    )
-                }
-            }
-            .fill(LinearGradient(
-                gradient: .init(colors: [Self.gradientStart, Self.gradientEnd]),
-                startPoint: .init(x: 0.5, y: 0),
-                endPoint: .init(x: 0.5, y: 0.6)
-            )) //図形の色をグラデーションにする
-                .aspectRatio(1, contentMode: .fit) //画面の中央に表示
-
+    static let rotationCount = 8 //表示するシンボルの数
+    
+    var badgeSymbols: some View {
+        ForEach(0..<Badge.rotationCount) { i in  //  0から7までの数を代入して角度を決める
+            RotatedBadgeSymbol(
+                angle: .degrees(Double(i) / Double(Badge.rotationCount)) * 360.0
+            )
         }
+            .opacity(0.5) //透過率
     }
-    static let gradientStart = Color(red: 239.0 / 255, green: 120.0 / 255, blue: 221.0 / 255)
-    static let gradientEnd = Color(red: 239.0 / 255, green: 172.0 / 255, blue: 120.0 / 255) //図形の色をグラデーションにする
-
+    
+    var body: some View {
+        
+        ZStack {
+            BadgeBackground() //バッジの背景画像
+            //            self.badgSymbol //回転するシンボル画像（背景画像の上に重なる）
+            
+            GeometryReader { geometry in
+                self.badgeSymbols //親ビューをもとに相対的に座標空間の値を返す//回転するシンボル画像
+                    .scaleEffect(1.0 / 4.0, anchor: .top) //大きさを縮小
+                    
+                    .position(x: geometry.size.width / 2.0, y: (3.0 / 4.0) * geometry.size.height) //画像の配置を指定
+            }
+        }
+        .scaledToFill()
+    }
 }
 
 struct Badge_Previews: PreviewProvider {
